backend/requirements.txt
Flask>=2.3.0
langchain==0.0.78
flask-cors==3.0.10
Werkzeug>=3.0.0
requests==2.28.2
python-dotenv==1.0.0
azure-identity
azure-keyvault-secrets
azure-storage-blob==12.19.0
Flask-Session==0.8.0
